#92 - Rename: 피드 폴더 이름 변경
